eval.py call metric core function to predict metric performance
